feat: COM-3512 Add getImagesLoaderRules
- fix code source
