+++ .storybook/manager.js
@@ -1,0 +1,6 @@
+import { addons } from '@storybook/addons';
+
+addons.setConfig({
+  showPanel: true,
+  panelPosition: 'right',
+});
