+++ .storybook/preview.js
@@ -1,6 +1,5 @@
 import React from 'react';
-import { Global } from '@emotion/react';
-import { ThemeProvider } from '@emotion/react';
+import { Global, ThemeProvider, css } from '@emotion/react';
 import { Root, Fonts, defaultTheme } from '@csssr/core-design';
 import { getTransformSource } from './utils/getTransformSource';
 
@@ -54,10 +53,6 @@ export const parameters = {
       const { args, kind } = storyContext;
       return getTransformSource(kind, args);
     },
-    inlineStories: false,
-    source: {
-      state: 'open',
-    },
   },
 };
 export const decorators = [
@@ -69,6 +64,10 @@ export const decorators = [
           <Story />
           <Global styles={stylesLayout} />
           <Global styles={(theme) => stylesBody({ theme })} />
+          {/* Для переопределения стилей блога, иначе код code source в одну строку */}
+          <Global styles={css`code {
+            white-space: pre;
+          }`}/>
         </Grid>
       </ThemeProvider>
     </Root>
